Add files via upload
This is a rough notebook consisting of all of the work that has been put into the backend resource. This IS NOT the final version of a the notebook, and will be updated as we go. The cleaned version will be added to the repo soon. Feel free to take a look but it might not make much sense at first as it it is, again, a VERY rough version. Lots of EDA, data cleaning, and testing of different methods to create the backend resource.